Notification

+++ backend/users/views.py
@@ -21,6 +21,8 @@ from .models import Comments
 from .serializers import CommentsSerializer
 from users.authentication import CookieJWTAuthentication
 from rest_framework.decorators import authentication_classes
+# from .models import Notification
+# from .serializers import NotificationSerializer
 User = get_user_model()
 # ============================
 # Authentication with user
@@ -276,4 +278,37 @@ class CommentViewSet(viewsets.ModelViewSet):
             obj.numComments = max(getattr(obj, 'numComments', 1) - 1, 0)
             obj.save(update_fields=['numComments'])
 
+        instance.delete()
+
+@authentication_classes([CookieJWTAuthentication])
+class NotificationViewSet():
+    queryset = Notification.objects.all()
+    serializer = NotificationSerializer()
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        return Notification.objects.filter(user=self.request.user).order_by('created_at')
+
+    def perform_create(self, serializer):
+        serializer.save(user=self.request.user)
+
+@authentication_classes([CookieJWTAuthentication])
+class MarkAsSeenViewSet():
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request):
+        Notification.objects.filter(user=request.user, seen = False).update(seen=True)
+        return response({'status: 200 OK'}, status=200)
+
+@authentication_classes([CookieJWTAuthentication])
+class NotificationDeleteViewSet():
+    permission_classes = [IsAuthenticated]
+    
+    def perform_destroy(self, instance):
+        user = self.request.user
+
+        if not (user.is_staff or user.is_superuser or instance.user == user):
+            raise PermissionDenied("Bạn chỉ có thể xóa thông báo của chính mình.")
+
+        # Xoá thông báo
         instance.delete()
